requirements.txt
langchain==0.3.12
langchain-openai==0.2.12
langgraph==0.2.59
matplotlib==3.10.0
mplsoccer==1.4.0
numpy==2.2.0
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
statsbombpy==1.14.0
streamlit==1.41.1
